// packages/podlet-js/src/models/pod-quadlet.ts
export interface PodQuadlet {
  Service?: ServiceQuadlet;
  Pod: {
// ...
    ExitPolicy?: string;
// ...
    Label?: Array<string>;
// ...
    PublishPort?: Array<string>;
// ...
  };
}